web/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
jinja2
python-multipart
requests
google-genai

